Store day state in Application
Updated Application.js to track day statate. This way, when setDay is called, it changes the day state, then the Application component renders and passes the new day to the DayList component. The DAyList component renders and passes props to the DayListItem component (it's children) causing the updates to the selected visual state.